+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.7.9] - 2025-05-
-- feat (WIP): multiple backends in subprojects;
-- 
+## [0.8.0] - 2025-05-
+- cleanup: `TestProject` and `TestProjectWriter`;
+- feat: multiple backends in subprojects;
+- doc: added usage hints;
+
 ## [0.7.9] - 2025-05-09 - unpublished
 - doc: Scala Native;
 - cleanup: use TaskProvider instead of Task more;